Fixed Tests
Refactor code to work with test cases
